real-time/producer/requirements.txt
kafka-python
pyshark